Adding ButtonIconComponent and new start button on map screen

+++ RpgTurno/BossRushDevApp/Custom/Component/Map/Stage/MapNodeBannerComponent.cs
@@ -87,7 +87,7 @@ public class MapNodeBannerComponent : FrameComponent
 
     private int ApplyBounce(int baseValue)
     {
-        var bounce = GlobalVariablesDto.GetBounceValue();
+        var bounce = (int)GlobalVariablesDto.GetBounceValue();
 
         return baseValue - bounce;
     }
