+++ RpgTurno/BossRushDevApp/Custom/Component/Play/Banners/SkillDetailsBannerComponent.cs
@@ -78,7 +78,7 @@ public class SkillDetailsBannerComponent : FrameComponent
 
     private int ApplyBounce(int baseValue)
     {
-        var bounce = GlobalVariablesDto.GetBounceValue();
+        var bounce = (int)GlobalVariablesDto.GetBounceValue();
         return baseValue - bounce;
     }
